Chen issue744

Fixes #691, #744

See merge request !357

+++ src/trapdoor/lib/sampling/dgsampling.cpp
@@ -37,15 +37,15 @@ namespace lbcrypto {
 	// As described in Figure 2 of https://eprint.iacr.org/2017/308.pdf
 
 	template <class Element>
-	void LatticeGaussSampUtility<Element>::GaussSampGq(const Element &syndrome, double stddev, size_t k, const typename Element::Integer &q, int32_t base,
-		typename Element::DggType &dgg, Matrix<int32_t> *z)
+	void LatticeGaussSampUtility<Element>::GaussSampGq(const Element &syndrome, double stddev, size_t k, const typename Element::Integer &q, int64_t base,
+		typename Element::DggType &dgg, Matrix<int64_t> *z)
 	{
 
 		// If DCRT is used, the polynomial is first converted from DCRT to large polynomial (in COEFFICIENT representation)		
 		typename Element::PolyLargeType u = syndrome.CRTInterpolate();
 
 		const typename Poly::Integer& modulus = u.GetParams()->GetModulus();
-		// std::cout << "modulus = " << modulus << std::endl; 
+		// std::cout << "modulus = " << modulus << std::endl;
 		double sigma = stddev / (base + 1);
 
 		// main diagonal of matrix L
@@ -69,17 +69,17 @@ namespace lbcrypto {
 
 		// c can be pre-computed as it only depends on the modulus
 		// (double) is added to avoid integer division
-		c(0, 0) = modulus.GetDigitAtIndexForBase(1, base) / (double)base;
+		c(0, 0) = (int64_t)(modulus.GetDigitAtIndexForBase(1, base)) / (double)base;
 
 		for (size_t i = 1; i < k; i++)
-			c(i, 0) = (c(i - 1, 0) + modulus.GetDigitAtIndexForBase(i + 1, base)) / base;
+			c(i, 0) = (c(i - 1, 0) + (int64_t)(modulus.GetDigitAtIndexForBase(i + 1, base))) / base;
 
 #pragma omp parallel for
 		for (size_t j = 0; j < u.GetLength(); j++)
 		{
 			typename Element::Integer v(u.at(j));
 
-			vector<int32_t> p(k);
+			vector<int64_t> p(k);
 
 			LatticeGaussSampUtility<Element>::Perturb(sigma, k, u.GetLength(), l, h, base, dgg, &p);
 
@@ -89,21 +89,21 @@ namespace lbcrypto {
 			// when the result is negative, a(0,0) gets values close to 2^32 if the cast is not used
 			//****a(0, 0) = ((int32_t)(v.GetDigitAtIndexForBase(1, base)) - p[0]) / base;
 			// (double) is added to avoid integer division
-			a(0, 0) = ((int32_t)(v.GetDigitAtIndexForBase(1, base)) - p[0]) / (double)base;
+			a(0, 0) = ((int64_t)(v.GetDigitAtIndexForBase(1, base)) - p[0]) / (double)base;
 
 			for (size_t t = 1; t < k; t++){
-				a(t, 0) = (a(t - 1, 0) + (int32_t)(v.GetDigitAtIndexForBase(t + 1, base)) - p[t]) / base;
+				a(t, 0) = (a(t - 1, 0) + (int64_t)(v.GetDigitAtIndexForBase(t + 1, base)) - p[t]) / base;
 			}
-			vector<int32_t> zj(k);
+			vector<int64_t> zj(k);
 
 			LatticeGaussSampUtility<Element>::SampleC(c, k, u.GetLength(), sigma, dgg, &a, &zj);
 
-			(*z)(0, j) = base*zj[0] + modulus.GetDigitAtIndexForBase(1, base)*zj[k - 1] + v.GetDigitAtIndexForBase(1, base);
+			(*z)(0, j) = base*zj[0] + (int64_t)(modulus.GetDigitAtIndexForBase(1, base))*zj[k - 1] + (int64_t)(v.GetDigitAtIndexForBase(1, base));
 
 			for (size_t t = 1; t < k - 1; t++){
-					(*z)(t, j) = base*zj[t] - zj[t - 1] + modulus.GetDigitAtIndexForBase(t + 1, base)*zj[k - 1] + v.GetDigitAtIndexForBase(t + 1, base);
+					(*z)(t, j) = base*zj[t] - zj[t - 1] + (int64_t)(modulus.GetDigitAtIndexForBase(t + 1, base))*zj[k - 1] + (int64_t)(v.GetDigitAtIndexForBase(t + 1, base));
 			}
-			(*z)(k - 1, j) = modulus.GetDigitAtIndexForBase(k, base)*zj[k - 1] - zj[k - 2] + v.GetDigitAtIndexForBase(k, base);
+			(*z)(k - 1, j) = (int64_t)(modulus.GetDigitAtIndexForBase(k, base))*zj[k - 1] - zj[k - 2] + (int64_t)(v.GetDigitAtIndexForBase(k, base));
 		}
 
 	}
@@ -113,15 +113,15 @@ namespace lbcrypto {
 	// As described in Algorithm 3 of https://eprint.iacr.org/2017/844.pdf
 
 	template <class Element>
-	void LatticeGaussSampUtility<Element>::GaussSampGqArbBase(const Element &syndrome, double stddev, size_t k, const typename Element::Integer &q, int32_t base,
-		typename Element::DggType &dgg, Matrix<int32_t> *z)
+	void LatticeGaussSampUtility<Element>::GaussSampGqArbBase(const Element &syndrome, double stddev, size_t k, const typename Element::Integer &q, int64_t base,
+		typename Element::DggType &dgg, Matrix<int64_t> *z)
 	{
 
 		// If DCRT is used, the polynomial is first converted from DCRT to large polynomial (in COEFFICIENT representation)		
 		typename Element::PolyLargeType u = syndrome.CRTInterpolate();
 
 		const typename Poly::Integer& modulus = u.GetParams()->GetModulus();
-		// std::cout << "modulus = " << modulus << std::endl; 
+		// std::cout << "modulus = " << modulus << std::endl;
 		double sigma = stddev / (base + 1);
 
 		// main diagonal of matrix L
@@ -145,10 +145,10 @@ namespace lbcrypto {
 
 		// c can be pre-computed as it only depends on the modulus
 		// (double) is added to avoid integer division
-		c(0, 0) = modulus.GetDigitAtIndexForBase(1, base) / (double)base;
+		c(0, 0) = ((int64_t)modulus.GetDigitAtIndexForBase(1, base) )/ (double)base;
 
 		for (size_t i = 1; i < k; i++)
-			c(i, 0) = (c(i - 1, 0) + modulus.GetDigitAtIndexForBase(i + 1, base)) / base;
+			c(i, 0) = (c(i - 1, 0) + (int64_t)(modulus.GetDigitAtIndexForBase(i + 1, base))) / base;
 
 #pragma omp parallel for
 		for (size_t j = 0; j < u.GetLength(); j++)
@@ -166,21 +166,21 @@ namespace lbcrypto {
 			//****a(0, 0) = ((int32_t)(v.GetDigitAtIndexForBase(1, base)) - p[0]) / base;
 			// (double) is added to avoid integer division
 
-			a(0, 0) = ((int32_t)(v.GetDigitAtIndexForBase(1, base)) - p[0]) / (double)base;
+			a(0, 0) = ((int64_t)(v.GetDigitAtIndexForBase(1, base)) - p[0]) / (double)base;
 
 			for (size_t t = 1; t < k; t++) {
-				a(t, 0) = (a(t - 1, 0) + (int32_t)(v.GetDigitAtIndexForBase(t + 1, base)) - p[t]) / base;
+				a(t, 0) = (a(t - 1, 0) + (int64_t)(v.GetDigitAtIndexForBase(t + 1, base)) - p[t]) / base;
 			}
-			vector<int32_t> zj(k);
+			vector<int64_t> zj(k);
 
 			LatticeGaussSampUtility<Element>::SampleC(c, k, u.GetLength(), sigma, dgg, &a, &zj);
 
-			(*z)(0, j) = base*zj[0] + modulus.GetDigitAtIndexForBase(1, base)*zj[k - 1] + v.GetDigitAtIndexForBase(1, base);
+			(*z)(0, j) = base*zj[0] + (int64_t)(modulus.GetDigitAtIndexForBase(1, base))*zj[k - 1] + (int64_t)(v.GetDigitAtIndexForBase(1, base));
 
 			for (size_t t = 1; t < k - 1; t++) {
-				(*z)(t, j) = base*zj[t] - zj[t - 1] + modulus.GetDigitAtIndexForBase(t + 1, base)*zj[k - 1] + v.GetDigitAtIndexForBase(t + 1, base);
+				(*z)(t, j) = base*zj[t] - zj[t - 1] + (int64_t)(modulus.GetDigitAtIndexForBase(t + 1, base))*zj[k - 1] + (int64_t)(v.GetDigitAtIndexForBase(t + 1, base));
 			}
-			(*z)(k - 1, j) = modulus.GetDigitAtIndexForBase(k, base)*zj[k - 1] - zj[k - 2] + v.GetDigitAtIndexForBase(k, base);
+			(*z)(k - 1, j) = (int64_t)(modulus.GetDigitAtIndexForBase(k, base))*zj[k - 1] - zj[k - 2] + (int64_t)(v.GetDigitAtIndexForBase(k, base));
 		}
 
 	}
@@ -192,7 +192,7 @@ namespace lbcrypto {
 
 	template <class Element>
 	void LatticeGaussSampUtility<Element>::Perturb(double sigma, size_t k, size_t n,
-		const vector<double> &l, const vector<double> &h, int32_t base, typename Element::DggType &dgg, vector<int32_t> *p) {
+		const vector<double> &l, const vector<double> &h, int64_t base, typename Element::DggType &dgg, vector<int64_t> *p) {
 
 		std::vector<int32_t> z(k);
 		double d = 0;
@@ -217,14 +217,14 @@ namespace lbcrypto {
 
 	template <class Element>
 	void LatticeGaussSampUtility<Element>::PerturbFloat(double sigma, size_t k, size_t n,
-		const vector<double> &l, const vector<double> &h, int32_t base, typename Element::DggType &dgg, vector<double> *p) {
+		const vector<double> &l, const vector<double> &h, int64_t base, typename Element::DggType &dgg, vector<double> *p) {
 
 		std::normal_distribution<> d(0, sigma);
 
 		std::mt19937 &g = PseudoRandomNumberGenerator::GetPRNG();
 
 		std::vector<double> z(k);
-		
+
 		// Generate a vector using continuous Gaussian distribution
 		for (size_t i = 0; i < k; i++)
 			z[i] = d(g);
@@ -243,19 +243,19 @@ namespace lbcrypto {
 
 	template <class Element>
 	void LatticeGaussSampUtility<Element>::SampleC(const Matrix<double> &c, size_t k, size_t n,
-		double sigma, typename Element::DggType &dgg, Matrix<double> *a, vector<int32_t> *z)
+		double sigma, typename Element::DggType &dgg, Matrix<double> *a, vector<int64_t> *z)
 	{
 		(*z)[k - 1] = dgg.GenerateIntegerKarney(-(*a)(k - 1, 0) / c(k - 1, 0), sigma / c(k - 1, 0));
 		//(*z)[k - 1] = dgg.GenerateInteger(-(*a)(k - 1, 0) / c(k - 1, 0), sigma / c(k - 1, 0),n);
 		*a = *a - ((double)((*z)[k - 1]))*c;
 
-		for (size_t i = 0; i < k - 1; i++) { 
+		for (size_t i = 0; i < k - 1; i++) {
 			(*z)[i] = dgg.GenerateIntegerKarney(-(*a)(i, 0), sigma);
 			//(*z)[i] = dgg.GenerateInteger(-(*a)(i, 0), sigma, n);
 		}
 
 	}
-	
+
 	// Subroutine used by ZSampleSigmaP as described Algorithm 4 in https://eprint.iacr.org/2017/844.pdf
 	// a - field element in DFT format
 	// b - field element in DFT format
@@ -274,7 +274,7 @@ namespace lbcrypto {
 
 			shared_ptr<Matrix<int64_t>> q2Int  = ZSampleF(dCoeff,c(1,0),dgg,n);
 			Field2n q2(*q2Int);
-			
+
 			Field2n q2Minusc2 = q2 - c(1, 0);
 			//Convert to DFT representation prior to multiplication
 			q2Minusc2.SwitchFormat();
@@ -282,7 +282,7 @@ namespace lbcrypto {
 			Field2n product = b * d.Inverse() * q2Minusc2;
 			//Convert the product to coefficient representation
 			product.SwitchFormat();
-			
+
 			//Computes c1 in coefficient format
 			Field2n c1 = c(0, 0) + product;
